FPL_Optimizer/requirements.txt
Flask==2.3.3
requests==2.31.0
pandas
numpy
pulp==2.7.0
streamlit
plotly
scikit-learn
azure-storage-blob
xgboost
python-dotenv
google-generativeai
